requirements.txt
channels==3.0.4
django-bootstrap-form==3.4
django-captcha==1.0.0
django-channels==0.7.0
django-cors-headers==3.8.0
django-emailit==0.2.4
django-gdpr-assist==1.3.0
django-newsletter==0.9.1
django-oscar-api==3.0.0
django-payments==0.14.0
django-photologue==3.13
django-simple-captcha==0.5.14
django-simpleui==2021.8.17
elasticsearch==7.14.1
pinax-announcements==4.0.0
pinax-badges==3.0.0
pinax-blog==8.0.1
pinax-calendars==3.0.0
pinax-eventlog==5.1.1
pinax-events==3.0.0
pinax-messages==3.0.0
pinax-testimonials==3.0.0
pip-chill==1.0.1
psycopg2==2.9.1
pycountry==20.7.3
pyopenssl==20.0.1
service-identity==21.1.0
whitenoise==5.3.0